requirements.txt
CherryPy==3.8.0
coverage==4.0
Jinja2==2.8
mock==1.0.1
mockldap==0.2.3
MarkupSafe==0.23
nose==1.3.4
python-ldap==2.4.20
simplejson==3.8.0
smbpasswd==1.0.2
Sphinx==1.3.1
sphinx-rtd-theme==0.1.9
wsgiref==0.1.2
